[eemm_AV] Add option to use common random numbers.
An option to use common c++11 random numbers was added. These can
be generated in parallel on the CPU while the GPU is being set up.

This features the macro MGONGPU_COMMONRAND_ONHOST, which is either
defined to true or false. This way, the compiler always checks both
code paths, but for the inactive path, no assembly is generated.