[cuda eemm] Allow for compiling cpu code without cuda.
To run simple compilation tests, it may be beneficial to compile even
when cuda is not installed on a system. This modifies the Makefile such
that if no nvcc is in the PATH, only the cpu-only part without cuda is
compiled.

Use this to switch off curand:
make MGONGPU_CONFIG=-DMGONGPU_COMMONRAND_ONHOST