[klas2ep12] epoch1 add '512z' build and rename all other simd builds
The idea is to distinguish between AVX512 on xmm/ymm and AVX512 on zmm registers.
The former ("AVX256" in LHCb) is the fastest option.
The latter is slower, probably because it causes a closk slowdown? (to be demonstrated)

Note in any case that '512z' seems to have many more instructions in total (why?)
=== Symbols in ./build.512y/CPPProcess.o === (~sse4:    0) (avx2: 2597) (512y:   94) (512z:    0)
=== Symbols in ./build.512z/CPPProcess.o === (~sse4:    0) (avx2: 1182) (512y:  208) (512z: 2035)