[ggttgg] ADDITIONAL BUG FIX related to issue#198 - FIX IXXXXX ALGORITHM AS IN FORTRAN
Note: issue #198 was about C++/CUDA giving different results and was fixed by the previous commit,
which only changed the behaviour of C++ while leaving CUDA unchanged.

This additional bug fix instead changes the behaviour of both CUDA and C++.
The results from the two implementations were the same before and are the same after the fix,
but in both backends the numerical results has changed.

This is a fix that I already applied in epoch1 eemumu for ixxxxx.
I changed the handling of pvec0 (energy) depending on the input fmass.
I used the same algorithm used in the Fortran.
To be cross checked by Olivier...

for exe in ./gcheck.exe ./check.exe; do echo; echo "$exe -v 1 8 1 | tail -10"; $exe -v 1 8 1 | tail -10; done

./gcheck.exe -v 1 8 1 | tail -10
Momenta:
   1  7.500000e+02  0.000000e+00  0.000000e+00  7.500000e+02
   2  7.500000e+02  0.000000e+00  0.000000e+00 -7.500000e+02
   3  3.118787e+02  1.091564e+02 -2.764329e+02  9.454122e+01
   4  2.317597e+02 -6.474121e+01 -1.333573e+02 -1.781487e+02
   5  3.853875e+02  1.954334e+00 -7.753632e+01  3.775021e+02
   6  5.709741e+02 -4.636953e+01  4.873266e+02 -2.938947e+02
--------------------------------------------------------------------------------
 Matrix element = 5.46032e-06 GeV^-4
--------------------------------------------------------------------------------

./check.exe -v 1 8 1 | tail -10
Momenta:
   1  7.500000e+02  0.000000e+00  0.000000e+00  7.500000e+02
   2  7.500000e+02  0.000000e+00  0.000000e+00 -7.500000e+02
   3  3.118787e+02  1.091564e+02 -2.764329e+02  9.454122e+01
   4  2.317597e+02 -6.474121e+01 -1.333573e+02 -1.781487e+02
   5  3.853875e+02  1.954334e+00 -7.753632e+01  3.775021e+02
   6  5.709741e+02 -4.636953e+01  4.873266e+02 -2.938947e+02
--------------------------------------------------------------------------------
 Matrix element = 5.46032e-06 GeV^-4
--------------------------------------------------------------------------------