[gtest] upgrade googletest from 1.10.0 to 1.11.0 This is needed to port to gcc11.1 and avoid [-Werror=maybe-uninitialized]
The code builds, but the test fails

[----------] 1 test from GoogleTestVerification
[ RUN      ] GoogleTestVerification.UninstantiatedParameterizedTestSuite<MadgraphTestFloat>
src/MadgraphTest.cc:166: Failure
Parameterized test suite MadgraphTestFloat is defined via TEST_P, but never instantiated. None of the test cases will run. Either no INSTANTIATE_TEST_SUITE_P is provided or the only ones provided expand to nothing.

Ideally, TEST_P definitions should only ever be included as part of binaries that intend to use them. (As opposed to, for example, being placed in a library that may be linked in to get other utilities.)

To suppress this error for this test suite, insert the following line (in a non-header) in the namespace it is defined in:

GTEST_ALLOW_UNINSTANTIATED_PARAMETERIZED_TEST(MadgraphTestFloat);
[  FAILED  ] GoogleTestVerification.UninstantiatedParameterizedTestSuite<MadgraphTestFloat> (0 ms)
[----------] 1 test from GoogleTestVerification (0 ms total)

[----------] Global test environment tear-down
[==========] 4 tests from 4 test suites ran. (733 ms total)
[  PASSED  ] 3 tests.
[  FAILED  ] 1 test, listed below:
[  FAILED  ] GoogleTestVerification.UninstantiatedParameterizedTestSuite<MadgraphTestFloat>