[fvsc] ** COMPLETE FVSC ** rerun all tmad/madX tests on the five processes after removing helicity filtering (#461) - much better
Difference between cudacpp ME performance in madevent and in SA/bridge is within 10% for complex processes.

For eemumu and ggtt the difference is still quite large instead.
This is now understood: madevent throughput is lower because it includes data copies/transposing
(in the SA "bridge" mode the timers really only compute the ME calculation)