Streamline cuda architecture flags.
To easily switch between architectures, streamline the setting of
architecture flags to a single line. These can be overridden by
exporting or setting MADGRAPH_CUDA_ARCHITECTURE.
By default, device code and PTX for architecture 70 will be embedded, so
it would run on Volta and (after a delay for JITting) on Ampere.