[fpe] in ggtt.sa testxxx.cc, add prepareTest (and improve FPE signal handler).
This also includes a resetHstMomentaToPar0, which is commented out for the moment.
The idea was to modify the momenta befaore each xxx call, to ensure that they are all consistent.
But I will instead implement a more solid fix.

No change in runTest behaviour, FPEs by default, succeeds if FPEs disabled