[fvsc] rerun tmad gguu: the xsec from fortran and cpp is the same with the new implementation of getgoodhel (#630 is fixed!)
This is still very surprising, because a ME sum of N differing from a sum of N+1 is an E-16 effect, while cross sections differe by a factor 2?

Also very strange that setting the top mass to 0 was allowing all helicities to be selected correctly?

Anyway, will backport to CODEGEN, and regenerate and retest all processes