[CODEGEN] Remove matrix1 counter from Fortran MEs.
Since matrix1 is called in a loop over all events, instrumenting it
with a timer and counter creates significant overhead.
Curiously, this counter is never read, so it can be removed.

For pp -> tt~ 012j, the difference for 1M events on 16 cores is:
Walltime: 35:48 m --> 7:45 m
CPU time: 13:14 h --> 2:56 h