[launch] in epochX/cudacpp, remove from .gitignore all the files in SubProcesses/P*/.gitignore
The main issue here is that this prevents adding bin/madevent to the repo.
This script is needed, for instance, to run ./bin/generate_events with a different runcard.
Presently this gives this error:
  ../bin/madevent treatcards run
  make: ../bin/madevent: No such file or directory

This is the complete list of files removed from cudacpp/.gitignore:
  .libs
  .cudacpplibs
  madevent
  madevent_fortran
  madevent_cpp
  madevent_cuda
  G[0-9]*
  ajob[0-9]*
  input_app.txt
  symfact.dat
  gensym

I will then regenerate all processes including bin/madevent