[pptt] in CODEGEN for CPPProcess.cc, add back the static_assert expecting nprocesses == 1 (part of the fix for #872 in PR #935; undo changes from PR #754 and #764)
NB: nprocesses is unused in cudacpp, because cudacpp is unable to handle mirror processes!
This means that the result of madevent with cudacpp MEs are the same whether nprocesses is 1 or 2.
However, nprocesses=2 signals that Fortran is using mirror processes.
If mirror processes are used in Fortran MEs but not in cudacpp MEs, this leads to a xsec mismatch #872.

Implementing mirror processes in cudacpp would be extremely complicated.
This is an assumption that was made years ago.
To be cross checked with Olivier, but I think that the code generated with cudacpp without mirror processes
also ensure that the mirror processes are _separately_ generated as separate subprocesses also in Fortran,
therefore the whole calculation is internally consistent if mirrorprocesses is false in both Fortran and cudacpp.