[hack_ihel3/cublas] in gg_tt.mad and CODEGEN, use CUBLAS_TF32_TENSOR_OP_MATH if env CUDACPP_RUNTIME_CUBLASTF32TENSOR is set
This forces the use of tensor cores for FPTYPE=m,f BLAS color sums on A100 GPUs and beyond
(but it has no effect on V100 GPUs)

It was tested on an A100 that, even in gg_tt.mad (ncolor=2) with only few events (-p 1 8 1),
tensor cores are used in gemmStridedBatched calls if this option is set, but they are not used otherwise.
(In particular, ampere_sgemm kernels which do not use tensor core use are launched if it is not set,
while cutlass kernels which use tensor cores are launched if it is set).

Profile tesor cores using ncu:
  CUDACPP_RUNTIME_BLASCOLORSUM=1 CUDACPP_RUNTIME_CUBLASTF32TENSOR=1 \
    ncu --metrics sm__pipe_tensor_cycles_active,sm__inst_executed_pipe_tensor \
    --target-processes all ./build.cuda_m_inl0_hrd0/check_cuda.exe -p 1 8 1