Raise samples difference threshold to 5%
Reference samples have been generated using MadtRex on top of vanilla
MadGraph and not on top of CUDACPP, so there is a slightly larger
uncertainty when comparing those ref samples with the reweighted samples
that start from CUDACPP-generated events.
We are not interested to check theat they are digit-by-digit correct,
but that they are almost correct with each other.