[hack_ihel6] ggttg.dpg5dpf5: in DCDIAG=1 avoid creating the wfs buffer
This completes the DCDIAG=1 implementation. Adding RDC has a performance penalty but it is not huge, around 10-20%.
NB: This is a comparison to split kernels that are retrieving wavefunctions from global memory.
This should be studied in more complex processes such as ggttgggg.

alias 'mscalingtest=for b in 1 2 4 8 16 32 64 128 256 512 1024; do ( ./build.cuda_m_inl0_hrd0/check_cuda.exe -p $b 256 1 \
  | \grep "EvtsPerSec\[MECalcOnly\]"
  | awk -vb=$b "{printf \"%s %4d %3d\n\", \$5, b, 256}" ) |& sed "s/Gpu.*Assert/Assert/"; done; unset CUDACPP_RUNTIME_BLASCOLORSUM'

DCDIAG=0
3.919665e+05    1 256
7.883533e+05    2 256
1.542838e+06    4 256
2.757143e+06    8 256
3.986899e+06   16 256
4.738626e+06   32 256
4.877144e+06   64 256
4.993746e+06  128 256
5.102723e+06  256 256
4.859062e+06  512 256
4.884247e+06 1024 256

DCDIAG=1
6.460633e+05    1 256
1.282774e+06    2 256
2.364972e+06    4 256
3.021224e+06    8 256
3.554574e+06   16 256
3.758733e+06   32 256
3.919725e+06   64 256
4.044548e+06  128 256
4.183804e+06  256 256
4.233798e+06  512 256
4.262484e+06 1024 256